+++ src/main/java/com/h5radar/radar/config/WebConfiguration.java
@@ -5,6 +5,8 @@ import org.springframework.context.annotation.Configuration;
 import org.springframework.web.servlet.config.annotation.InterceptorRegistry;
 import org.springframework.web.servlet.config.annotation.WebMvcConfigurer;
 
+import com.h5radar.radar.domain.AuthRequestInterceptor;
+
 @Configuration
 public class WebConfiguration implements WebMvcConfigurer {
